fix: add @muhlemmer suggested changes
