+++ cmd/start/start.go
@@ -13,6 +13,8 @@ import (
 	"time"
 
 	clockpkg "github.com/benbjohnson/clock"
+	"github.com/common-nighthawk/go-figure"
+	"github.com/fatih/color"
 	"github.com/gorilla/mux"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -22,6 +24,7 @@ import (
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
 
+	"github.com/zitadel/zitadel/cmd/build"
 	"github.com/zitadel/zitadel/cmd/key"
 	cmd_tls "github.com/zitadel/zitadel/cmd/tls"
 	"github.com/zitadel/zitadel/internal/actions"
@@ -111,6 +114,8 @@ type Server struct {
 }
 
 func startZitadel(config *Config, masterKey string, server chan<- *Server) error {
+	showBasicInformation(config)
+
 	ctx := context.Background()
 
 	dbClient, err := database.Connect(config.Database, false)
@@ -453,4 +458,30 @@ func shutdownServer(ctx context.Context, server *http.Server) error {
 	}
 	logging.New().Info("server shutdown gracefully")
 	return nil
+}
+
+func showBasicInformation(startConfig *Config) {
+	fmt.Println(color.MagentaString(figure.NewFigure("Zitadel", "", true).String()))
+	http := "http"
+	if startConfig.TLS.Enabled || startConfig.ExternalSecure {
+		http = "https"
+	}
+
+	consoleURL := fmt.Sprintf("%s://%s:%v/ui/console\n", http, startConfig.ExternalDomain, startConfig.ExternalPort)
+	healthCheckURL := fmt.Sprintf("%s://%s:%v/debug/healthz\n", http, startConfig.ExternalDomain, startConfig.ExternalPort)
+
+	insecure := !startConfig.TLS.Enabled && !startConfig.ExternalSecure
+
+	fmt.Printf(" ===============================================================\n\n")
+	fmt.Printf(" Version          : %s\n", build.Version())
+	fmt.Printf(" TLS enabled      : %v\n", startConfig.TLS.Enabled)
+	fmt.Printf(" External Secure  : %v\n", startConfig.ExternalSecure)
+	fmt.Printf(" Console URL      : %s", color.BlueString(consoleURL))
+	fmt.Printf(" Health Check URL : %s", color.BlueString(healthCheckURL))
+	if insecure {
+		fmt.Printf("\n %s: you're using plain http without TLS. Be aware this is \n", color.RedString("Warning"))
+		fmt.Printf(" not a secure setup and should only be used for test systems.         \n")
+		fmt.Printf(" Visit: %s    \n", color.CyanString("https://zitadel.com/docs/self-hosting/manage/tls_modes"))
+	}
+	fmt.Printf("\n ===============================================================\n\n")
 }
